[webgpu-headers] Changes counts to use size_t in webgpu headers.
- Updates code generators to handle size_t appropriately on the wire.
- Fixes OOB read on wire test.
- Updates dawn node conversion types.

Bug: github.com/webgpu-native/webgpu-headers/issues/172
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/135623
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Loko Kung <lokokung@google.com>

+++ src/dawn/native/PipelineLayout.cpp
@@ -20,6 +20,7 @@
 
 #include "dawn/common/Assert.h"
 #include "dawn/common/BitSetIterator.h"
+#include "dawn/common/Numeric.h"
 #include "dawn/common/ityp_stack_vec.h"
 #include "dawn/native/BindGroupLayout.h"
 #include "dawn/native/Device.h"
@@ -61,7 +62,8 @@ PipelineLayoutBase::PipelineLayoutBase(DeviceBase* device,
                                        ApiObjectBase::UntrackedByDeviceTag tag)
     : ApiObjectBase(device, descriptor->label) {
     ASSERT(descriptor->bindGroupLayoutCount <= kMaxBindGroups);
-    for (BindGroupIndex group(0); group < BindGroupIndex(descriptor->bindGroupLayoutCount);
+    for (BindGroupIndex group(0);
+         group < BindGroupIndex(checked_cast<uint32_t>(descriptor->bindGroupLayoutCount));
          ++group) {
         mBindGroupLayouts[group] = descriptor->bindGroupLayouts[static_cast<uint32_t>(group)];
         mMask.set(group);
