+++ src/dawn/node/binding/GPURenderBundleEncoder.cpp
@@ -48,7 +48,7 @@ void GPURenderBundleEncoder::setBindGroup(
 
     wgpu::BindGroup bg{};
     uint32_t* offsets = nullptr;
-    uint32_t num_offsets = 0;
+    size_t num_offsets = 0;
     if (!conv(bg, bindGroup) || !conv(offsets, num_offsets, dynamicOffsets)) {
         return;
     }
